refactor: add cache to a new module
