+++ faster_sam/middlewares/lambda_authorizer.py
@@ -14,7 +14,8 @@ from starlette.requests import Request
 from starlette.responses import Response
 from starlette.types import ASGIApp
 
-from faster_sam import web_identity_providers, cache
+from faster_sam import web_identity_providers
+from faster_sam.cache.interface import CacheInterface
 
 
 logger = logging.getLogger(__name__)
@@ -208,7 +209,7 @@ class LambdaAuthorizerMiddleware(BaseHTTPMiddleware):
         app: ASGIApp,
         lambda_function: str,
         credentials: Credentials = Credentials(),
-        cache: str = None,
+        cache: CacheInterface = None,
     ) -> None:
         """
         Initializes the LambdaAuthorizer.
@@ -250,26 +251,21 @@ class LambdaAuthorizerMiddleware(BaseHTTPMiddleware):
         Response
             The response generated by the middleware.
         """
+        if self._cache is None:
+            payload = self.invoke_lambda(request)
+        else:
+            key = self.get_key(dict(request.headers))
 
-        payload = None
-
-        if self._cache is not None:
-            redis_cache = cache.factory(self._cache)
-            redis_cache.set_key(dict(request.headers))
-
-            if redis_cache.token is None:
+            if key is None:
                 content = {"message": "Unauthorized"}
                 status_code = HTTPStatus.UNAUTHORIZED
                 return Response(content=json.dumps(content), status_code=status_code.value)
 
-            payload = redis_cache.get()
+            payload = self._cache.get(key)
 
             if payload is None:
                 payload = self.invoke_lambda(request)
-                redis_cache.set(payload, 900)
-
-        if payload is None:
-            payload = self.invoke_lambda(request)
+                self._cache.set(key, payload)
 
         if payload and payload["policyDocument"]["Statement"][0]["Effect"] == "Allow":
             request.scope["authorization_context"] = payload.get("context")
@@ -352,3 +348,11 @@ class LambdaAuthorizerMiddleware(BaseHTTPMiddleware):
             },
         }
         return event
+
+    def get_key(self, headers: dict) -> str:
+        headers = {key.lower(): value for key, value in headers.items()}
+
+        if "AUTHORIZATION" in os.environ:
+            return os.environ["AUTHORIZATION"]
+
+        return headers.get("authorization", headers.get("x-api-key"))
