+++ pyproject.toml
@@ -23,3 +23,4 @@ dependencies = [
 
 [project.optional-dependencies]
 auth = ["boto3", "requests"]
+redis = ["redis"]
